requirements.txt
numpy
opencv-python
scikit-image
scikit-learn
joblib
customtkinter
pillow
